% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sex.R
\name{get_sex}
\alias{get_sex}
\title{Guess Sex from a personal name}
\usage{
get_sex(x, threshold = 0.5)
}
\arguments{
\item{x}{A character vector of personal names.}

\item{threshold}{Minimum probability to guess sex. If real probability is lower, the function returns NA.}
}
\value{
Either "male", or "female", or NA.
}
\description{
Guess Sex from a personal name
}
