% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.cze-package.R
\docType{package}
\name{names.cze-package}
\alias{names.cze}
\alias{names.cze-package}
\title{names.cze: Get Information on Personal Names Used in the Czech Republic}
\description{
Frequency table of personal names in the Czech Republic.
    Guess type (forename, surname) or sex (male, female) of any name.
}
\keyword{internal}
